tests/wpt/tests/html/browsers/history/the-location-interface/location_href.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_href</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    test(function () {
      var href = location.href;

      assert_equals(href, document.URL, "href");

    }, "location href");
    </script>
  </body>
</html>
